refactor/begins the process of collecting code into modules
