+++ index.html
@@ -25,6 +25,6 @@
       <button id="game__start-btn" class="game__start-btn">Start</button>
     </main>
 
-    <script src="./src/main.js"></script>
+    <script src="./src/main.js" type="module"></script>
   </body>
 </html>
